app/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
supabase>=2.0,<3.0
python-jose[cryptography]>=3.3,<4.0
httpx>=0.27,<1.0
anthropic>=0.25,<1.0
google-genai>=1.0,<2.0
openai>=1.12,<2.0
groq>=0.5,<1.0
stripe>=8.0,<10.0
PyMuPDF>=1.24,<2.0
Pillow>=10.0,<12.0
python-multipart>=0.0.9,<1.0
python-docx>=1.1,<2.0
tjd-db-handler>=0.1.2,<1.0

